feat: BufferPoolSharded 实现 IBufferPool 接口
- 添加接口头文件引用
- 类声明继承 IBufferPool
- 统一 PageId 类型为 int32_t

+++ src/storage_engine/buffer_pool/buffer_pool_interface.h
@@ -9,7 +9,8 @@ namespace sqlcc {
 namespace storage {
 
 class Page;
-using PageId = uint64_t;
+// PageId 保持与现有实现一致
+using PageId = int32_t;
 
 /**
  * IBufferPool - 缓冲区管理接口
